requirements.txt
absl-py==1.4.0
addict==2.4.0
attrs==23.1.0
brotlipy==0.7.0
cachetools==5.3.0
certifi
cffi
charset-normalizer
click==8.1.3
cryptography
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.1
faiss-gpu==1.7.2
filelock==3.9.0
flit_core
fonttools==4.38.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
idna
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mmcls==0.23.2
mmcv-full==1.6.1
mmsegmentation==0.27.0
numpy
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.3.0
prettytable==3.4.1
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2023.5.5
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
sentry-sdk==1.15.0
setproctitle==1.2.2
shortuuid==1.0.8
six
smmap==5.0.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.6.11
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions
urllib3
wandb==0.13.2
wcwidth==0.2.6
Werkzeug==2.2.3
yapf==0.32.0
zipp==3.15.0
